Add scripts to install on GKE Cluster. Buf Fixes and enhanced documentation

+++ install/08-uninstall-all.sh
@@ -1,18 +1,24 @@
 #!/bin/bash
 
-set -e
 SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &> /dev/null && pwd)
-cd "$SCRIPT_DIR"
+cd "$SCRIPT_DIR" || true
+
+[ ! -f "$SCRIPT_DIR"/../.env ] && echo "Error: Environment file missing, exiting." 1>&2 && exit 1
+set -o allexport; source "$SCRIPT_DIR"/../.env; set +o allexport
 
 WORKLOAD_NAME=${1:-$DEFAULT_WORKLOAD_NAME}
 
+TAP_INSTALL_NS_STATUS=$(kubectl get ns tap-install --ignore-not-found=true -ojson | jq .status.phase -r)
+TAP_CLUSTER_ESSENTIALS_NS_STATUS=$(kubectl get ns tanzu-cluster-essentials --ignore-not-found=true -ojson | jq .status.phase -r)
+
+
 echo "========================================Removing TAP Packages========================================"
-tanzu package installed list \
+[ "$TAP_INSTALL_NS_STATUS" == "Active" ] && tanzu package installed list \
   --namespace tap-install \
   -o json | jq -c '.[].name' | xargs -n1 tanzu package installed delete --yes --namespace tap-install
 
 echo "========================================Removing TAP Repositories========================================"
-tanzu package repository list \
+[ "$TAP_INSTALL_NS_STATUS" == "Active" ] && tanzu package repository list \
    --namespace \
    tap-install \
    -o json | jq -c '.[].name' | xargs -n1 tanzu package repository delete --yes --namespace tap-install
@@ -20,16 +26,16 @@ tanzu package repository list \
 echo "========================================Removing Tanzu Cluster Essentials========================================"
 if [ -d "$HOME"/tanzu-cluster-essentials ]; then
   cd "$HOME"/tanzu-cluster-essentials || exit
-  ./kapp delete -a secretgen-controller -n tanzu-cluster-essentials --yes || true
-  ./kapp delete -a kapp-controller -n tanzu-cluster-essentials --yes || true
+  [ "$TAP_CLUSTER_ESSENTIALS_NS_STATUS" == "Active" ] && { ./kapp delete -a secretgen-controller -n tanzu-cluster-essentials --yes || true; }
+  [ "$TAP_CLUSTER_ESSENTIALS_NS_STATUS" == "Active" ] && { ./kapp delete -a kapp-controller -n tanzu-cluster-essentials --yes || true; }
 fi
-kubectl delete ns tanzu-cluster-essentials || true
+[ "$TAP_CLUSTER_ESSENTIALS_NS_STATUS" == "Active" ] && { kubectl delete ns tanzu-cluster-essentials || true; }
 rm -rf "$HOME"/tanzu-cluster-essentials || true
 rm -rf /usr/local/bin/kapp || true
 
 echo "========================================Removing Tanzu CLI========================================"
 rm -rf "$HOME"/tanzu/cli || true
-rm /usr/local/bin/tanzu || true
+rm -rf /usr/local/bin/tanzu || true
 rm -rf "$HOME"/.config/tanzu || true
 rm -rf "$HOME"/.tanzu || true
 rm -rf "$HOME"/tanzu || true
@@ -37,16 +43,10 @@ rm -rf "$HOME"/.cache/tanzu || true
 rm -rf "$HOME"/Library/Application\ Support/tanzu-cli/* || true
 
 echo "========================================Removing TAP Namespaces========================================"
-kubectl delete ns tap-install || true
-cd "$SCRIPT_DIR"
+[ "$TAP_INSTALL_NS_STATUS" == "Active" ] && { kubectl delete ns tap-install || true; }
+cd "$SCRIPT_DIR" || true
 if [ -d generated ]; then rm -rd generated; fi
 
-
-if [[ $1 == 'gcloud' ]]; then
-   rm_gcr "$WORKLOAD_CONTAINER_REGISTRY"/build-service
-   rm_gcr "$WORKLOAD_CONTAINER_REGISTRY"/supply-chain/"$WORKLOAD_NAME"-"$DEVELOPER_NAMESPACE"-bundle
-   rm_gcr "$WORKLOAD_CONTAINER_REGISTRY"/supply-chain/"$WORKLOAD_NAME"-"$DEVELOPER_NAMESPACE"
-fi
 
 rm_gcr(){
  local C=0
@@ -61,3 +61,15 @@ rm_gcr(){
   done
   echo "Deleted ${C} images in ${IMAGE}." >&2
 }
+
+if [ "$K8_CLUSTER" == 'gke' ]; then
+
+   echo "Deleting GKE Cluster..."
+   gcloud container clusters delete tap-cluster -q --region="$GCP_REGION"
+
+   echo "Deleting container images..."
+   rm_gcr "$CONTAINER_REGISTRY_HOSTNAME"/"$CONTAINER_REPOSITORY"/build-service
+   rm_gcr "$CONTAINER_REGISTRY_HOSTNAME"/"$CONTAINER_REPOSITORY"/supply-chain/"$WORKLOAD_NAME"-"$DEVELOPER_NAMESPACE"-bundle
+   rm_gcr "$CONTAINER_REGISTRY_HOSTNAME"/"$CONTAINER_REPOSITORY"/supply-chain/"$WORKLOAD_NAME"-"$DEVELOPER_NAMESPACE"
+fi
+
